Move open_rd_file to its own file
